% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbanalysis.r
\name{summary.pb}
\alias{summary.pb}
\title{Summary method}
\usage{
\method{summary}{pb}(obj)
}
\arguments{
\item{obj}{A pb object generated by the pb.fit function}
}
\description{
Summary method
}
\examples{
#get summary for ordinal


data = data.frame(bmi_cat = NHANES::NHANES$BMI_WHO,
                  race = NHANES::NHANES$Race1,
                  age = NHANES::NHANES$Age,
                  poverty = NHANES::NHANES$Poverty
)
data = na.omit(data)

out = pb.fit(bmi_cat ~ age + poverty,
             data = data,
             disparity.group = "race",
             majority.group = "White",
             minority.group = c("Black"),
             prop.odds.fail = c("age"),
             family = "ordinal")

summary(out)
}
